requirements.txt
asgiref==3.4.1
Django==2.0.7
django-bootstrap-toolkit==2.15.0
django-crispy-forms==1.13.0
django-filter==21.1
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
